+++ app/models/user.rb
@@ -9,10 +9,19 @@ class User < ActiveRecord::Base
   belongs_to  :role
 
   def role
-    Role.find(role_id).role.to_sym if role_id
+    @role ||= Role.find(role_id).role.to_sym if role_id
   end
   def role? role
     role = Role.find_by_role(role.to_s)
     role_id == role.try(:id)
+  end
+  def admin?
+    role === :admin
+  end
+  def manager?
+    role === :manager
+  end
+  def user?
+    role === :user
   end
 end